Move DNS provider initialization to DNS controller
* cmd/ingress-operator/start.go (NewStartCommand): Delete initialization of
the DNS provider.
(cloudCredentialsSecretName, createDNSProvider, getPlatformStatus): Move
from here...
* pkg/operator/controller/dns/controller.go: ...to here.
(New): Delete parameter for DNS provider; the controller now initializes
it.  Add config parameter.  Add watches on the DNS config and
Infrastructure config objects.  Use ToDNSRecords mapper function to
reconcile all DNSRecord objects when a config object changes.
(Config): New type to hold the namespace and operator release version,
which are needed for initializing the DNS provider.
(reconciler): Add Config and infraConfig fields.
(Reconcile): Get the cluster Infrastructure config object, and initialize
or re-initialize the DNS provider if it was not already initialized or if
the Infrastructure config has changed.
(ToDNSRecords): New method.  Return reconcilation requests for all
DNSRecords.
* pkg/operator/operator.go (New): Update initialization of the DNS
controller to pass the controller's config object instead of the DNS
provider object.
* manifests/00-cluster-role.yaml: Allow the operator to list and watch the
dnses and infrastructures config resources.
* pkg/manifests/bindata.go: Regenerate.
