+++ manifests/00-cluster-role.yaml
@@ -86,6 +86,8 @@ rules:
 - apiGroups:
   - config.openshift.io
   resources:
+  - dnses
+  - infrastructures
   - ingresses
   verbs:
   - list
