+++ pkg/manifests/bindata.go
@@ -11,7 +11,7 @@
 // assets/router/service-account.yaml (213B)
 // assets/router/service-cloud.yaml (631B)
 // assets/router/service-internal.yaml (429B)
-// manifests/00-cluster-role.yaml (2.411kB)
+// manifests/00-cluster-role.yaml (2.441kB)
 // manifests/00-custom-resource-definition-internal.yaml (6.371kB)
 // manifests/00-custom-resource-definition.yaml (42.103kB)
 // manifests/00-ingress-credentials-request.yaml (1.464kB)
@@ -316,7 +316,7 @@ func assetsRouterServiceInternalYaml() (*asset, error) {
 	return a, nil
 }
 
-var _manifests00ClusterRoleYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xbc\x96\xcf\x8e\xdb\x38\x0c\xc6\xef\x7e\x0a\x61\x72\x5b\x20\x0e\xf6\xb6\xc8\x6d\xb1\x0b\xf4\xd4\x0e\x50\x14\xbd\xcb\x12\x1d\xb3\x91\x45\x81\xa4\x92\xa6\x4f\x5f\xf8\xdf\x64\x26\xf1\x24\x6e\x3b\xed\x29\x16\x4d\x7d\xfc\x49\xfc\x24\x67\x65\xfe\x0b\x59\x14\xd8\x30\x05\x30\x35\xb1\xd1\x06\x0c\x25\x60\xab\xc4\x06\x55\x20\xd4\x65\xb1\x32\x9f\x1e\xff\x7f\xdc\x9a\x7f\x4d\x20\x35\x54\x77\x59\x02\x46\x1a\xca\xc1\x9b\x0a\x0c\x43\x0a\xd6\x81\x37\xd5\xa9\x97\x12\x83\xb1\x97\x8a\xb6\x05\x49\xd6\x81\xf4\xea\xc7\x06\x5d\x53\xac\x5e\x56\xb1\x4e\xb3\x0d\xe1\x64\x22\x80\x17\x63\x9d\x03\x91\xb2\xd8\x63\xf4\xdb\x09\xf0\x23\x05\x28\x6c\xc2\xcf\xc0\x82\x14\xb7\x86\x2b\xeb\x4a\x9b\xb5\x21\xc6\x6f\x56\x91\x62\xb9\xff\x47\x4a\xa4\xcd\xe1\xef\xa2\x05\xb5\xde\xaa\xdd\x16\xa6\x27\xd8\x76\xc5\xa2\x34\x58\xeb\x1a\xe3\x8e\x41\x64\x3d\x95\x2f\x38\x07\x90\x6d\xb1\x36\x36\xe1\x3b\xa6\x9c\xa4\x9b\xb6\x36\x0f\x0f\x85\x31\x0c\x42\x99\x1d\x8c\x31\x47\xb1\xc6\x5d\x6b\x93\xf4\xc3\xf3\xea\xfa\xa1\x00\x1f\xd0\x81\x75\x8e\x72\xd4\x21\x06\xd1\x27\xc2\x69\x34\x66\x4c\x03\xc7\x30\xbe\x48\xe4\xc7\xfc\x03\x0c\xc9\x07\xe0\x6a\x22\xf9\xeb\xa1\xb8\xe6\xb3\xa9\xa7\xb8\x20\xf4\x90\x02\x9d\xda\xa5\x22\x89\x02\xba\xd3\xb5\x4c\x22\xef\x51\x38\xa7\x6e\x67\xab\xec\x77\xb0\x4c\xaf\xa5\x88\x4a\x8c\x71\x57\x3a\x62\x20\x29\x1d\xb5\xd7\xf2\xe3\x3e\x8c\xd9\x17\xca\x8e\xc1\x2a\xf4\x8f\x3b\xd0\xfe\x37\x27\xdf\x85\xae\xeb\xbd\x6a\x83\x99\xde\x0d\x4e\xea\xed\x79\x19\xa8\x30\x7a\x8c\xbb\x21\x7e\xce\xb8\x78\x75\x9b\x31\xa0\x0c\x0f\x47\xab\xae\x99\x61\x9d\x1c\x57\x3e\x99\x71\x96\x73\x34\xa8\xa3\xa8\x4c\x21\x00\xcb\x2b\xe1\x8d\xa8\xd5\xbc\xa8\x2d\xe3\xe4\x72\x21\x82\x8f\xc2\xe0\x88\x47\x4f\x9e\x87\x3f\x50\x72\x38\x2a\x77\xd7\x5a\xb3\x15\xe5\xec\x34\x33\xbc\x58\x28\x3c\xd5\x1e\x9f\x6c\xc2\xce\x36\xd3\x7e\x44\xd0\x23\xf1\xfe\x82\xa5\x6b\xc6\x4f\xb3\x9c\xeb\x3e\x53\xbc\xdb\xd6\x45\xe2\xa3\xd7\xa6\xfd\xbf\xed\xa6\x37\xaa\x31\xdb\xac\xe9\x24\xad\xcc\x7b\x64\x26\x06\x6f\x6a\xa6\xd6\x58\x11\x50\xd9\x30\x65\x05\xde\xb4\xa0\x8c\x4e\x36\xa3\xe4\xba\x3b\x08\xe5\xc9\xb6\x61\xe6\x04\x76\x33\xee\x90\x0d\xaa\x32\xc9\xce\xb4\xec\x36\xce\x02\x8c\xee\x0e\x80\xa8\xe8\x6e\x5f\x02\x4a\x7b\x88\x0c\x07\x84\xe3\x7c\x0f\xde\x86\xe4\xfe\x6d\x24\xb9\xfa\x02\x4e\x87\x8f\xdd\xef\x05\x5a\xdc\xa1\x3b\xce\xff\x83\x20\xbf\x60\xdd\x05\x1c\x02\x2e\x33\xea\xe9\x0e\xca\x94\xd6\x5d\xb9\xf0\x55\x1d\x45\x51\xb6\x78\xf5\x65\xcd\x02\xcf\x26\x7f\xe8\xfe\x13\x0c\x2f\x1a\x12\x1d\xaf\xa9\xe2\x7b\x00\x00\x00\xff\xff\x37\x23\x0f\x74\x6b\x09\x00\x00")
+var _manifests00ClusterRoleYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xbc\x96\xcf\x8e\xdb\x38\x0c\xc6\xef\x7e\x0a\x61\x72\x5b\x20\x0e\xf6\xb6\xc8\x6d\xb1\x0b\xf4\xd4\x0e\x50\x14\xbd\x2b\x12\x1d\xb3\x91\x45\x81\xa4\x92\xa6\x4f\x5f\xf8\xdf\x64\x26\xf1\xc4\x6e\x3b\xed\x29\x16\x4d\x7d\xfc\x49\xfc\x24\x67\x65\xfe\x0b\x59\x14\xd8\x30\x05\x30\x15\xb1\xd1\x1a\x0c\x25\x60\xab\xc4\x06\x55\x20\x54\x65\xb1\x32\x9f\x1e\xff\x7f\xdc\x9a\x7f\x4d\x20\x35\x54\xb5\x59\x02\x46\x6a\xca\xc1\x9b\x1d\x18\x86\x14\xac\x03\x6f\x76\xe7\x4e\x4a\x0c\xc6\x4e\x2a\xda\x06\x24\x59\x07\xd2\xa9\x9f\x6a\x74\x75\xb1\x7a\x59\xc5\x3a\xcd\x36\x84\xb3\x89\x00\x5e\x8c\x75\x0e\x44\xca\xe2\x80\xd1\x6f\x47\xc0\x8f\x14\xa0\xb0\x09\x3f\x03\x0b\x52\xdc\x1a\xde\x59\x57\xda\xac\x35\x31\x7e\xb3\x8a\x14\xcb\xc3\x3f\x52\x22\x6d\x8e\x7f\x17\x0d\xa8\xf5\x56\xed\xb6\x30\x1d\xc1\xb6\x2d\x16\xa5\xc6\x4a\xd7\x18\xf7\x0c\x22\xeb\xb1\x7c\xc1\x39\x80\x6c\x8b\xb5\xb1\x09\xdf\x31\xe5\x24\xed\xb4\xb5\x79\x78\x28\x8c\x61\x10\xca\xec\x60\x88\x39\x8a\x15\xee\x1b\x9b\xa4\x1b\x5e\x56\xd7\x0d\x05\xf8\x88\x0e\xac\x73\x94\xa3\xf6\x31\x88\x3e\x11\x8e\xa3\x21\x63\x1c\x38\x86\xe1\x45\x22\x3f\xe4\x1f\xa1\x4f\x3e\x02\xef\x46\x92\xbf\x1e\x8a\x5b\x3e\x9b\x3a\x8a\x2b\x42\x0f\x29\xd0\xb9\x59\x2a\x92\x28\xa0\x3b\xdf\xca\x24\xf2\x1e\x85\x73\x6a\x77\x76\x97\xfd\x1e\x96\xe9\x35\x14\x51\x89\x31\xee\x4b\x47\x0c\x24\xa5\xa3\xe6\x56\x7e\xd8\x87\x21\xfb\x4a\xd9\x31\x58\x85\xee\x71\x0f\xda\xfd\xe6\xe4\xdb\xd0\x6d\xbd\x57\x6d\x30\xd1\xbb\xde\x49\x9d\x3d\xaf\x03\x3b\x8c\x1e\xe3\xbe\x8f\x5f\x32\xae\x5e\xdd\x67\x0c\x28\xfd\xc3\xc9\xaa\xab\x27\x58\x47\xc7\x95\x4f\x66\x9c\xe4\x1c\x0c\xea\x28\x2a\x53\x08\xc0\xf2\x4a\x78\x23\x6a\x35\x2f\x6a\xcb\x30\xb9\x5c\x88\xe0\xa3\x30\x38\xe2\xc1\x93\x97\xe1\x0f\x94\xec\x8f\xca\xec\x5a\x2b\xb6\xa2\x9c\x9d\x66\x86\x17\x0b\x85\xa7\xda\xc3\x93\x4d\xd8\xda\x66\xdc\x8f\x08\x7a\x22\x3e\x5c\xb1\xb4\xcd\xf8\x49\x96\x4b\xa5\x39\xaa\x67\xf5\x66\x9b\xbe\xa8\xf4\xe0\xc4\xb1\x3b\xf7\xbd\xf6\x46\x35\x26\x5b\x39\x9e\xb3\x95\x79\x8f\xcc\xc4\xe0\x4d\xc5\xd4\x18\x2b\x02\x2a\x1b\xa6\xac\xc0\x9b\x06\x94\xd1\xc9\x66\x90\x5c\xb7\xc7\xa4\x3c\xdb\x26\x4c\x9c\xcf\x76\xc6\x0c\x59\xaf\x2a\xa3\xec\x44\x43\xef\xe3\x2c\xc0\x68\x6f\x08\x88\x8a\xee\xfe\x15\xa1\x74\x80\xc8\x70\x44\x38\x4d\xf7\xe0\x6d\x48\xe6\xef\x2a\xc9\xbb\x2f\xe0\xb4\xff\x14\xfe\x5e\xa0\xc5\x1d\x9a\x71\xfe\x1f\x04\xf9\x05\xeb\x2e\xe0\x10\x70\x99\x51\xcf\x33\x28\x63\x5a\x7b\x21\xc3\x57\x75\x14\x45\xd9\xe2\xcd\x77\x37\x0b\x3c\x9b\xfc\xa1\xfd\xc7\xd0\xbf\xa8\x49\x74\xb8\xc4\x8a\xef\x01\x00\x00\xff\xff\x8c\x57\x39\x8e\x89\x09\x00\x00")
 
 func manifests00ClusterRoleYamlBytes() ([]byte, error) {
 	return bindataRead(
@@ -331,8 +331,8 @@ func manifests00ClusterRoleYaml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "manifests/00-cluster-role.yaml", size: 2411, mode: os.FileMode(420), modTime: time.Unix(1, 0)}
-	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x5e, 0x59, 0xbf, 0x86, 0x16, 0xa4, 0xe7, 0xa3, 0xa8, 0xb7, 0xfb, 0x93, 0x93, 0x19, 0x47, 0x9d, 0xe2, 0x50, 0xe, 0xe3, 0x6a, 0x20, 0x29, 0x75, 0x76, 0x36, 0x35, 0x9d, 0x58, 0x6c, 0x36, 0x4a}}
+	info := bindataFileInfo{name: "manifests/00-cluster-role.yaml", size: 2441, mode: os.FileMode(420), modTime: time.Unix(1, 0)}
+	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0x6e, 0xd6, 0xda, 0xe, 0x3d, 0x67, 0xbd, 0x76, 0xc2, 0x7f, 0xc0, 0x77, 0x94, 0xe6, 0xe9, 0x82, 0xe9, 0x79, 0x8b, 0x23, 0xef, 0x9a, 0xed, 0xe3, 0x32, 0x4e, 0xaf, 0x81, 0xd9, 0xcb, 0xbf, 0xe6}}
 	return a, nil
 }
